requirements.txt
pyfiglet
pyyaml
datetime
selenium
wikipedia
wolframalpha
requests
feedparser
bs4
speechrecognition
aiohttp
flask
Flask-SQLAlchemy
pyaudio
fpdf



